+++ 03-frontend/angular-ecommerce/src/app/components/product-details/product-details.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { Product } from 'src/app/common/product';
 import { ProductService } from 'src/app/services/product.service';
 import { ActivatedRoute } from '@angular/router';
+import { CartService } from 'src/app/services/cart.service';
+import { CartItem } from 'src/app/common/cart-item';
 
 @Component({
   selector: 'app-product-details',
@@ -12,8 +14,10 @@ export class ProductDetailsComponent implements OnInit {
 
   // fixes race conditions
   product: Product = new Product();
+  
   constructor(private productService: ProductService, 
-               private route: ActivatedRoute) { }
+              private cartService: CartService,
+              private route: ActivatedRoute) { }
 
   ngOnInit(): void {
     this.route.paramMap.subscribe(() => {
@@ -33,4 +37,10 @@ export class ProductDetailsComponent implements OnInit {
     )
   }
 
+  addToCart() {
+
+    console.log(`adding to cart: ${this.product.name}, ${this.product.unitPrice}`);
+    const theCartItem = new CartItem(this.product);
+    this.cartService.addToCart(theCartItem);
+  }
 }